NERDTreeToggle instead of NERDTreeFind on ,n
Using :NERDTreeFind when having just opened vim sets the root directory
to that above $PWD, since it navigates to the current file. Most of the
time navigating to a file is only desired when the current file is a
regular file (not a directory), so switch to using NERDTreeToggle
instead -- I don't think I need to navigate to the current file in the
tree very often.