Re-add profile name, rm job name from iTerm2 tabs
The profile name is the terminal title (i.e. tab name), so it can be set
by escape sequence (`\033]0;new title\a`), which seems handy now that I
have the tab names on the Touch Bar.

The job name takes up too much space and it generally not very useful,
so remove that.