Add script for updating Go project dependencies
I often run a series of dep commands to check the status of and update
vendored dependencies, then I copypaste the status output via some
temporary file into the message of a commit that records the update.

This script automates that. Run it from the root of a project. It tries
to update dependencies and creates a commit with a proper message if it
succeeds.

The resulting commit looks something like this:

  commit 4c966fad6acbf4a1f7962dcc406a941c94726a5e (HEAD -> stage)
  Author: Foo Bar <foo@bar>
  Date:   Fri May 18 17:46:41 2018 +0200

      Update vendored dependencies

      $ dep status && dep ensure && dep ensure -update && echo && dep status
      PROJECT                CONSTRAINT     VERSION        REVISION  LATEST   PKGS USED
      github.com/beevik/ntp  v0.2.0         v0.2.0         62c80a0   v0.2.0   1
      golang.org/x/net       branch master  branch master  6078986   2491c5d  4

      PROJECT                CONSTRAINT     VERSION        REVISION  LATEST   PKGS USED
      github.com/beevik/ntp  v0.2.0         v0.2.0         62c80a0   v0.2.0   1
      golang.org/x/net       branch master  branch master  2491c5d   2491c5d  4