Set VAULT_ADDR port explicitly to avoid SRV lookup
The vault cli is terribly slow; 'vault status' takes around 30 seconds
to run. This seems to be related to how DNS resolving works in Go on
macOS. Compiling the vault binary from source instead of using the
binary provided by HashiCorp or the one from Homebrew in combination
with setting the Vault server port explicitly seems to solve it.

  brew uninstall vault
  go get github.com/hashicorp/vault
  export VAULT_ADDR=https://vault.b17g.services:443/

Details:

  deverton commented on 16 Jul 2018
    My guess is the delay is caused by this line of code
    https://github.com/hashicorp/vault/blob/838a449c9b48/api/client.go#L560
    which does a SRV lookup if the port is not specified.

    @nickmaccarthy can you try specifying your Vault address including the
    port, e.g. :443 and see if its quicker?

  nickmaccarthy commented on 16 Jul 2018
    Hi @deverton , that worked! Specifying :443 at the end of VAULT_ADDR
    solved it. Thank you, I had worked around this by building a linux VM
    with vault. Thank you for finding that!

https://github.com/hashicorp/vault/issues/3709#issuecomment-405144597