git: Use diff-highlight for better highlighting
This sets core.pager as mentioned by diff-highlight's readme, and also
interactive.diffFilter to get the same highlighting for add -p and
friends.

scripts/diff-highlight/{README,diff-highlight} copied from
/usr/local/share/git-core/contrib/diff-highlight/ as of Git 2.28.0 as
installed by brew on macOS. The path to this contrib script is different
on Raspbian and the script is not even executable, so copying seems like
a sane enough choice.