iTerm2: Use native full-screen windows
The OS menu bar sometimes isn't accessible from non-native full-screen
and it's a bit of a hassle to get to the desktop (reverse pinch-zoom, or
switch to non-full-screen mode). In native full-screen mode, I can swipe
to the side. There's still the animation when switching, but maybe I can
get used to it.