Revert "vim: Remove goimports fmt command override"
gopls often gets the line numbers wrong. I'm not sure why but I'm
guessing that it could be due to edits happening outside of the editor.
Using gopls for formatting in this scenario causes it to join and even
delete lines. Not acceptable. goimports worked without issues, so let's
continue using that.

This reverts commit dd29ad63a8853da4e3c198d6c2d376b2bc6be858.