tmux: Increase max width of status-right
If the status text is too long it gets cut off. This should fix that. I
can't find any downside to setting this to a high value. The default
value is 40. I'm setting it to 100 just to make it not cut off. It could
probably be set to a greater value if needed.